move wallet tests into integration module in advance of further test creation, add v3 API
